Added Interactable Script
Added an Interactable Script that acts as the superclass to all interactable objects.  The interactable object in this case is just a test phone script. do not actually use the test phone script to add functionality.
Currently the interactable script will check if the player is hovering over an interactable object and turn it yellow if there is. on a  "e" press that object's interact() method will be called. To add a new interactable object with the script, remove the monobehavior inheritence and replace it with Interactable parent class then remove the update function and add a function with the header
protected override void name(){}
Comments will be made in a future push. One small bug i found is that if you run the scene without maximizing the game scene the camera's direction will be offset with the interact ray's direction. not sure what is causing this