Added scripts for both the trashcan and bell
Note: this does not currently function as the real and fake pizzas were not defined yet. Also there are no audio clips yet.
    /*Precondition: the player must be named "Player"
     *              player must have the Orders_Note attached to it
     *Postcondition: plays a noice whether a real or fake pizza was submitted. When using the scripts make sure:
Precondition:
the player must be named "Player"
player must have the Orders_Note attached to it
Postcondition: plays a noise whether a real or fake pizza was submitted